requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
sqlalchemy>=2.0.30
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9
